ATMsecurityBackend/facialRecognition/requirements.txt
amqp
asgiref
attrs
autobahn
Automat
billiard
celery
certifi
cffi
channels
chardet
charset-normalizer
click
click-didyoumean
click-plugins
click-repl
colorama
constantly
cron-descriptor
cryptography
daphne
diff-match-patch
dj-rest-auth
Django
django-allauth
django-celery-beat
django-cors-headers
django-filter
django-import-export
django-jazzmin
django-phonenumber-field
django-redis
django-timezone-field
djangorestframework
djangorestframework_simplejwt
dlib
dnspython
eventlet
geographiclib
geopy
greenlet
hyperlink
idna
incremental
kombu
numpy
opencv-python-headless
phonenumbers
pillow
prompt_toolkit
psycopg2
pyasn1
pyasn1_modules
pycparser
PyJWT
PyMySQL
pyOpenSSL
PyPDF2
python-crontab
python-dateutil
python-decouple
redis
reportlab
requests
service-identity
setuptools
six
sqlparse
tablib
Twisted
txaio
typing_extensions
tzdata
urllib3
vine
wcwidth
whitenoise
zope.event
zope.interface
